modulize(?) dist class
